Calculate network fee with rpc server
